// src/types/data.types.ts
export type DataType = {
    id: number;
    title: string;
    description: string;
    price: number;
    discountPercentage: number;
    rating: number;
    stock: number;
    brand: string;
    category: string;
    thumbnail: string;
    images: string[];
}
export type DataPropsType = {
    data: DataType
    fetchData: () => void
}